+++ Material/pzmathyperelastic.h
@@ -39,7 +39,7 @@ char *Name() { return "TPZMatHyperElastic"; }
 virtual void Contribute(TPZVec<REAL> &x,TPZFMatrix &jacinv ,TPZVec<REAL> &sol,TPZFMatrix &dsol,REAL weight,
 			  TPZFMatrix &axes,TPZFMatrix &phi,TPZFMatrix &dphi,TPZFMatrix &ek,TPZFMatrix &ef);
 
-virtual void ContributeBC(TPZVec<REAL> &x,TPZVec<REAL> &sol,double weight,
+virtual void ContributeBC(TPZVec<REAL> &x,TPZVec<REAL> &sol,REAL weight,
 			    TPZFMatrix &axes,TPZFMatrix &phi,TPZFMatrix &ek,TPZFMatrix &ef,TPZBndCond &bc);
 
 #ifdef _AUTODIFF
